Killed avanced patching; directly extend React.(Pure)Component or dispose manually
The amount of messing with prototypes and hackery, didn't justify the maintenance of `patch` anymore, since it is no longer used for `observer`. Especially since there are still subtle edges that didn't work. So the new implementation made the usage very strict, since there is a clean escape hatch: do it manually as done before this utility existed